pyproject.toml
[tool.poetry]
name = "guru"
version = "0.1.0"
description = ""
authors = ["Matthew Rosen matthewhrosen@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^0.57.0"
dj-rest-auth = {version = "^5.0.2", extras = ["with_social"]}
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
